+++ R/provider-github.R
@@ -1,20 +1,19 @@
 #' Chat with a model hosted on the GitHub model marketplace
 #'
 #' @description
-#' GitHub (via Azure) hosts a number of open source and OpenAI models.
-#' To access the GitHub model marketplace, you will need to apply for and
-#' be accepted into the beta access program. See
-#' <https://github.com/marketplace/models> for details.
+#' GitHub Models hosts a number of open source and OpenAI models. To access the
+#' GitHub model marketplace, you will need to apply for and be accepted into the
+#' beta access program. See <https://github.com/marketplace/models> for details.
 #'
 #' This function is a lightweight wrapper around [chat_openai()] with
-#' the defaults tweaked for the GitHub model marketplace.
+#' the defaults tweaked for the GitHub Models marketplace.
+#'
+#' GitHub also suports the Azure AI Inference SDK, which you can use by setting
+#' `base_url` to `"https://models.inference.ai.azure.com/"`. This endpoint was
+#' used in \pkg{ellmer} v0.3.0 and earlier.
 #'
 #' @family chatbots
-#' @param api_key The API key to use for authentication. You generally should
-#'   not supply this directly, but instead manage your GitHub credentials
-#'   as described in <https://usethis.r-lib.org/articles/git-credentials.html>.
-#'   For headless environments, this will also look in the `GITHUB_PAT`
-#'   env var.
+#' @param api_key `r api_key_param("GITHUB_PAT")`
 #' @param model `r param_model("gpt-4o")`
 #' @export
 #' @inheritParams chat_openai
@@ -26,7 +25,7 @@
 #' }
 chat_github <- function(
   system_prompt = NULL,
-  base_url = "https://models.inference.ai.azure.com/",
+  base_url = "https://models.github.ai/inference/",
   api_key = github_key(),
   model = NULL,
   seed = NULL,
@@ -36,7 +35,7 @@ chat_github <- function(
 ) {
   check_installed("gitcreds")
 
-  model <- set_default(model, "gpt-4o")
+  model <- set_default(model, "gpt-4.1")
   echo <- check_echo(echo)
 
   chat_openai(
@@ -66,7 +65,7 @@ github_key <- function() {
 #' @rdname chat_github
 #' @export
 models_github <- function(
-  base_url = "https://models.inference.ai.azure.com/",
+  base_url = "https://models.github.ai/",
   api_key = github_key()
 ) {
   provider <- ProviderOpenAI(
@@ -77,20 +76,44 @@ models_github <- function(
   )
 
   req <- base_request(provider)
-  req <- req_url_path_append(req, "/models")
+  req <- req_url_path_append(req, "/catalog/models")
   resp <- req_perform(req)
 
   json <- resp_body_json(resp)
 
-  id <- map_chr(json, "[[", "name")
+  if (grepl("models.inference.ai.azure.com", base_url, fixed = TRUE)) {
+    # Support listing models from the older Azure endpoint (ellmer <= 0.3.0)
+    id <- map_chr(json, "[[", "name")
+    publisher <- map_chr(json, "[[", "publisher")
+    license <- map_chr(json, "[[", "license")
+    task <- map_chr(json, "[[", "task")
+
+    res <- data.frame(
+      id = id,
+      publisher = publisher,
+      license = license,
+      task = task
+    )
+    return(res)
+  }
+
+  id <- map_chr(json, "[[", "id")
   publisher <- map_chr(json, "[[", "publisher")
-  license <- map_chr(json, "[[", "license")
-  task <- map_chr(json, "[[", "task")
+  registry <- map_chr(json, "[[", "registry")
+  rate_limit_tier <- map_chr(json, "[[", "rate_limit_tier")
+  version <- map_chr(json, "[[", "version")
+  capabilities <- map_chr(
+    map(json, "[[", "capabilities"),
+    paste,
+    collapse = ", "
+  )
 
   data.frame(
     id = id,
     publisher = publisher,
-    license = license,
-    task = task
+    registry = registry,
+    rate_limit_tier = rate_limit_tier,
+    version = version,
+    capabilities = capabilities
   )
 }