+++ tests/testthat/test-provider-any.R
@@ -1,7 +1,7 @@
 test_that("useful errors", {
   expect_snapshot(error = TRUE, {
     chat()
-    chat("a/b/c")
+    chat("")
     chat("susan")
     chat("susan/jones")
   })
@@ -16,4 +16,9 @@ test_that("can set model or use default", {
   chat2 <- chat("openai/gpt-4.1-mini")
   expect_equal(chat2$get_provider()@name, "OpenAI")
   expect_equal(chat2$get_provider()@model, "gpt-4.1-mini")
+
+  # We split at the first slash, the remainder is passed to `model`
+  chat3 <- chat("openai/provider/model")
+  expect_equal(chat3$get_provider()@name, "OpenAI")
+  expect_equal(chat3$get_provider()@model, "provider/model")
 })